refactor(Packages): more repackaging and renames
