+++ app/src/main/java/com/example/recipereviews/utils/ImageUtils.java
@@ -17,7 +17,7 @@ import com.google.android.gms.common.util.Strings;
 import com.squareup.picasso.Picasso;
 import com.squareup.picasso.Target;
 
-public class ImageUtil {
+public class ImageUtils {
 
     public static void loadImage(ImageView imageView, String imageName) {
         if (!Strings.isEmptyOrWhitespace(imageName)) {
